Move parquet_writer.py to the parquet module
