+++ dae/dae/impala_storage/schema1/impala_schema1.py
@@ -6,7 +6,7 @@ import toml
 from dae.configuration.study_config_builder import StudyConfigBuilder
 from dae.utils import fs_utils
 from dae.impala_storage.schema1.import_commons import save_study_config
-from dae.import_tools.parquet_writer import ParquetWriter
+from dae.parquet.parquet_writer import ParquetWriter
 from dae.import_tools.import_tools import ImportStorage
 from dae.task_graph.graph import TaskGraph
 from dae.parquet.schema1.parquet_io import NoPartitionDescriptor, \
